# docs/openapi.yaml
openapi: 3.1.0
info:
  title: TTD API
  description: 코딩테스트 문제풀이 + AI 코드 해설 SaaS API 문서
  version: v0.0.1
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/subscriptions:
    post:
      tags:
      - subscription-controller
      operationId: subscribe
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionSubscribeRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSubscriptionResponse"
      security:
      - bearerAuth: []
    delete:
      tags:
      - subscription-controller
      operationId: cancel
      responses:
        "204":
          description: No Content
      security:
      - bearerAuth: []
  /api/subscriptions/me:
    get:
      tags:
      - subscription-controller
      operationId: getMySubscription
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSubscriptionResponse"
      security:
      - bearerAuth: []
components:
  schemas:
...
    SubscriptionSubscribeRequest:
      type: object
      properties:
        billingKey:
          type: string
          minLength: 1
      required:
      - billingKey
    SubscriptionResponse:
      type: object
      properties:
        status:
          type: string
          enum:
          - ACTIVE
          - PAST_DUE
          - CANCELED
          - EXPIRED
        currentPeriodStart:
          type: string
          format: date-time
        nextBillingAt:
          type: string
          format: date-time
        canceledAt:
          type: string
          format: date-time
    ApiResponseSubscriptionResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/SubscriptionResponse"
        errorCode:
          type: string
        message:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT